requirement.txt
python-twitter==3.5
urllib3==1.25.8
wordcloud==1.6.0
matplotlib==3.1.3
numpy==1.18.1
bs4==0.0.1